+++ activerecord/lib/active_record/base.rb
@@ -11,6 +11,7 @@ require "active_record/relation/delegation"
 require "active_record/attributes"
 require "active_record/type_caster"
 require "active_record/database_configurations"
+require "active_record/pattern_matching"
 
 module ActiveRecord # :nodoc:
   # = Active Record
@@ -328,6 +329,7 @@ module ActiveRecord # :nodoc:
     include SignedId
     include Suppressor
     include Encryption::EncryptableRecord
+    include PatternMatching::Record
   end
 
   ActiveSupport.run_load_hooks(:active_record, Base)