+++ activerecord/lib/active_record/relation.rb
@@ -17,6 +17,7 @@ module ActiveRecord
 
     include Enumerable
     include FinderMethods, Calculations, SpawnMethods, QueryMethods, Batches, Explain, Delegation
+    include PatternMatching::Relation
 
     attr_reader :table, :klass, :loaded, :predicate_builder
     attr_accessor :skip_preloading_value